+++ src/adcirc/geometry/element.h
@@ -27,7 +27,6 @@
 #define ELEMENT_H
 
 #include "adcirc/adcircmodules_global.h"
-#include "adcirc/geometry/element.h"
 #include "adcirc/geometry/node.h"
 #include <string>
 #include <vector>
@@ -38,10 +37,10 @@ namespace Geometry {
 class Element {
 public:
   explicit Element();
-  explicit Element(int id, Adcirc::Geometry::Node *n1,
+  explicit Element(size_t id, Adcirc::Geometry::Node *n1,
                    Adcirc::Geometry::Node *n2, Adcirc::Geometry::Node *n3);
 
-  void setElement(int id, Adcirc::Geometry::Node *n1,
+  void setElement(size_t id, Adcirc::Geometry::Node *n1,
                   Adcirc::Geometry::Node *n2, Adcirc::Geometry::Node *n3);
 
   int n() const;
@@ -49,15 +48,15 @@ public:
   Adcirc::Geometry::Node *node(int i);
   void setNode(int i, Adcirc::Geometry::Node *node);
 
-  int id() const;
-  void setId(int id);
+  size_t id() const;
+  void setId(size_t id);
 
   std::string toString();
 
 private:
   int m_n = 3;
 
-  int m_id;
+  size_t m_id;
   std::vector<Adcirc::Geometry::Node *> m_nodes;
 };
 }
