+++ src/adcirc/modelparameters/attributemetadata.h
@@ -31,7 +31,7 @@ namespace ModelParameters {
 class AttributeMetadata {
  public:
   explicit AttributeMetadata(string name = "defaultName", string units = "none",
-                             int numValues = 1);
+                             size_t numValues = 1);
 
   string name() const;
   void setName(const string &name);
@@ -39,13 +39,13 @@ class AttributeMetadata {
   string units() const;
   void setUnits(const string &units);
 
-  int numberOfValues() const;
-  void setNumberOfValues(int numValues);
+  size_t numberOfValues() const;
+  void setNumberOfValues(size_t numValues);
 
   vector<double> getDefaultValues() const;
-  double getDefaultValue(int index = 0) const;
+  double getDefaultValue(size_t index = 0) const;
   void setDefaultValue(const double &value);
-  void setDefaultValue(int index, const double &value);
+  void setDefaultValue(size_t index, const double &value);
   void setDefaultValue(const vector<double> &value);
 
  private:
@@ -56,7 +56,7 @@ class AttributeMetadata {
   string m_units;
 
   /// Number of values associated with each node for this nodal attribute
-  int m_numValues;
+  size_t m_numValues;
 
   /// Default value(s) to be used when writing the fort.13 file
   vector<double> m_defaultValue;
